fix: primitive mode serialized (#702)
* fix: MeshPrimitive.Mode is omitted from serialization when at its spec default.

Introduce an internal `ModeSerialized` shim property bound to the JSON "mode" field: it returns null when Mode is Triangles (so the framework omits it) and falls back to Triangles on deserialize when absent. The
public Mode property keeps its non-nullable type and Triangles default, so no consumer sites need to change.

Adds a JsonSerializationTests fixture covering the round-trip.

* fix: avoid serializing Sampler.WrapS/WrapT and BufferView.Target at defaults

BufferView.Target is now BufferViewTarget instead of int.

Adds JsonSerializationTests cases covering Sampler/BufferView defaults and per-value write paths, and JsonDeserializationTests cases covering the corresponding read paths (omitted JSON falls back to Repeat / Undefined; explicit values round-trip).

+++ Packages/com.unity.cloud.gltfast/Runtime/Scripts/Export/GltfWriter.cs
@@ -2361,7 +2361,7 @@ namespace GLTFast.Export
                 Buffer = 0,
                 ByteOffset = (int)byteOffset,
                 ByteLength = bufferViewData.Length,
-                Target = (int)bufferViewTarget,
+                Target = bufferViewTarget,
             };
             if (byteStride.HasValue)
             {
